scripts/requirements.txt
pyFirmata
opencv-python
l298n-pyfirmata
pandas
